本次更新部分基础架构，更新spring boot版本并引入spring security进行权限校验，同时使用jwt进行单点登录检查，日志方面排除spring boot默认的logback使用slf4j+log4j2进行日志记录

+++ src/main/java/com/huang/service/impl/WsServiceImpl.java
@@ -19,7 +19,7 @@ public class WsServiceImpl implements WsService {
     @Async
     @Override
     public void notifyMsgToUser(Long toUserId) {
-        int count = userMessageService.count(new QueryWrapper<MUserMessage>()
+        long count = userMessageService.count(new QueryWrapper<MUserMessage>()
                 .eq("to_user_id", toUserId)
                 .eq("status", "0")
         );
